% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nansenbiomass-package.R
\docType{package}
\name{nansenbiomass-package}
\alias{nansenbiomass}
\alias{nansenbiomass-package}
\title{nansenbiomass: survey biomass reconstruction with StoX and sdmTMB}
\description{
See \code{docs/spec.md} in the source repository for the specification,
including the data-protection model (Section 3) and the module layout
(Section 4).
}
\author{
\strong{Maintainer}: Nikolaos Nikolioudakis \email{nikolaos.nikolioudakis@hi.no}

Authors:
\itemize{
  \item Nikolaos Nikolioudakis \email{nikolaos.nikolioudakis@hi.no}
}

}
\keyword{internal}
